[REFACT] moved most helper syscall functions to file_operations; updated README documentation on how to allow other users to access the file system
