+++ Makefile
@@ -1,6 +1,6 @@
 VERSION = 0
 SUBVERSION = 2
-PATCH = 17
+PATCH = 18
 EXTRAVERSION = -a
 
 DEBUG = 0
